Update aws-ssm-ec2-proxy-command.sh
I could not make the original script work with the simple command expansion `$_` and worse it failed silently (and as I am admin I happened to have the root private key for all the instances on my ssh-add keychain so we could not figure out why devs could not access it but I could). 

I changed it to this to a less elegant but working solution for ubuntu/centos/fedora etc but obviously there might need to be an alternative solution for distros that do not use `/home/` for user directories.